requirements.txt
# Core ML/DL libraries
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Data processing
datasets>=2.14.0
transformers>=4.30.0
tokenizers>=0.13.0
numpy>=1.24.0
pandas>=2.0.0

# NLP and evaluation
nltk>=3.8.0
rouge-score>=0.1.2
sacrebleu>=2.3.1

# Text preprocessing
spacy>=3.6.0
sentencepiece>=0.1.99

# Reinforcement Learning
gym>=0.26.0

# Utilities
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0
wandb>=0.15.0
tensorboard>=2.13.0

# Configuration
pyyaml>=6.0
omegaconf>=2.3.0

# File handling
jsonlines>=3.1.0
h5py>=3.9.0

# Development
jupyter>=1.0.0
ipywidgets>=8.0.0
black>=23.0.0
flake8>=6.0.0
pytest>=7.4.0
PyPDF2>=3.0.0
pdfminer.six>=20240706
sentence-transformers>=2.2.2
ebooklib>=0.18
beautifulsoup4>=4.12.2
bert-score>=0.3.13